+++ src/Exceptions/VoucherRedeemedException.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace FrittenKeeZ\Vouchers\Exceptions;
 
-class VoucherAlreadyRedeemedException extends VoucherException
+class VoucherRedeemedException extends VoucherStateException
 {
     /**
      * Exception message.
@@ -12,11 +12,4 @@ class VoucherAlreadyRedeemedException extends VoucherException
      * @var string
      */
     protected $message = 'Voucher has already been redeemed.';
-
-    /**
-     * Exception code - we use 409 Conflict.
-     *
-     * @var int
-     */
-    protected $code = 409;
 }