+++ src/VouchersServiceProvider.php
@@ -9,22 +9,14 @@ use Illuminate\Support\ServiceProvider;
 class VouchersServiceProvider extends ServiceProvider
 {
     /**
-     * Indicates if loading of the provider is deferred.
-     *
-     * @var bool
-     */
-    protected $defer = false;
-
-    /**
      * Bootstrap the application services.
      */
     public function boot(): void
     {
         $this->publishes([$this->getPublishConfigPath() => config_path('vouchers.php')], 'config');
 
-        with(method_exists($this, 'publishesMigrations') ? 'publishesMigrations' : 'publishes', function ($method) {
-            $this->{$method}([$this->getPublishMigrationsPath() => database_path('migrations')], 'migrations');
-        });
+        $method = method_exists($this, 'publishesMigrations') ? 'publishesMigrations' : 'publishes';
+        $this->{$method}([$this->getPublishMigrationsPath() => database_path('migrations')], 'migrations');
     }
 
     /**